OCR/requirements.txt
# PyTorch с поддержкой CUDA 12.4
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0

# PaddlePaddle для PaddleOCR-VL
paddlepaddle-gpu==2.6.2

# Transformers из основной ветки (требуется для GLM-OCR)
git+https://github.com/huggingface/transformers.git@main

# FastAPI и зависимости
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.9

# Обработка изображений и PDF
pillow==10.4.0
PyMuPDF==1.24.2
opencv-python==4.10.0.84

# Прочие зависимости
accelerate==0.34.0
sentencepiece==0.2.0
numpy==1.26.0
einops==0.8.0
paddlenlp==2.8.0
requests==2.32.0